+++ src/App.tsx
@@ -1,9 +1,9 @@
 import { Suspense } from "react";
-import { ToastOptions, Toaster } from "react-hot-toast";
 
 import { Global } from "@emotion/react";
 
 import Login from "@components/Login";
+import Toaster from "@components/Toaster";
 
 import { getGlobalStyles } from "@styles/globalStyles";
 
@@ -14,22 +14,13 @@ import AppRouter from "./router";
 function App() {
   const { theme } = useThemeStore();
 
-  const defaultToasterOption: ToastOptions = {
-    duration: 2000,
-    style: {
-      backgroundColor: theme.BACKGROUND200,
-      color: theme.TEXT600,
-      boxShadow: theme.SHADOW
-    }
-  };
-
   return (
     <>
       <Global styles={getGlobalStyles(theme)} />
       <Suspense fallback={null}>
         <Login />
         <AppRouter />
-        <Toaster toastOptions={defaultToasterOption} />
+        <Toaster />
       </Suspense>
     </>
   );
